refactor(delta): split rabin.rs into directory module, remove unused extend_match variants
Split the 1520-line rabin.rs monolithic file into a focused directory module:

- rabin/mod.rs (1083 lines): core delta generation, public API, integration tests
- rabin/tables.rs (160 lines): precomputed T/U lookup tables and constants
- rabin/extend_match.rs (262 lines): word-at-a-time match extension and unit tests

Removed 4 unused extend_match profiling variants (indexed, sliced_iter,
sliced_index, ptr) — only the word-at-a-time variant remains, which benchmarks
show is consistently faster. Improved module-level documentation with algorithm
overview, phase descriptions, and edge-case notes.
